Change build language
[MAKEFILE#DELETED] <- Deleted Makefile.
[CMAKELISTS.TXT] <- Changed build language to Cmake, and therefore added
a `CMakeLists.txt`. It has rules for everything, and it does the exact
same thing that normal make does, just easier. For example, this is how
to add an executable from a C file in Make:
```Makefile
CC=<c_compiler>
install:
    $(CC) -o <bin> <src>
```
And this is how it is in CMake:
```
add_executable(<bin> <src>)
```
Much easier! At least for C projects (which is what Vine is, duh...).