Rewrite "configure" file in shell script
[configure.py->configure]:
Rewrite "configure" file in shell script. Why did I do this? Well, I've
been purging my projects of python (like FMake, if you've heard of it)
because python is trash. So I rewrote the file in a better language for
these purposes; shell script.
[README.rst]:
Update to align to changes in project.