Add Ctrl+D key to delete a whole row.
[src/vine.c]:
Add Ctrl+D (vine reads it as (d) & 0x1f) as key to delete a whole line
(kind of like "dd" in nvim), though it's kind of buggy, I got a segfault
because of it, but when I wanted to replicate it, I didn't get the
segfault! So yeah, use at your own risk.